changed data source to Data4All

+++ shared_types/src/scheduling_environment/work_order/operation/operation_analytic.rs
@@ -4,17 +4,17 @@ use serde::Serialize;
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct OperationAnalytic {
     pub preparation_time: f64,
-    pub duration: u32,
+    pub duration: f64,
 }
 
 impl OperationAnalytic {
-    pub fn new(preparation_time: f64, duration: u32) -> Self {
+    pub fn new(preparation_time: f64, duration: f64) -> Self {
         OperationAnalytic {
             preparation_time,
             duration,
         }
     }
-    pub fn duration(&self) -> u32 {
+    pub fn duration(&self) -> f64 {
         self.duration
     }
 }
